requirements.txt
wget https://huggingface.co/TheBloke/Mistral-7B-Instruct-v0.1-GGUF/resolve/main/mistral-7b-instruct-v0.1.Q5_K_M.gguf
faiss-cpu==1.9.0
fastapi==0.115.0
filelock==3.16.1
filetype==1.2.0
frozenlist==1.4.1
fsspec==2024.9.0
greenlet==3.1.1
huggingface-hub==0.25.2
langchain==0.3.3
langchain-community==0.3.2
langchain-core==0.3.10
langchain-text-splitters==0.3.0
langdetect==1.0.9
langsmith==0.1.132
llama_cpp_python==0.3.1
lxml==5.3.0
nltk==3.9.1
numpy==1.26.4
openpyxl==3.1.5
pandas==2.2.3
pillow==10.4.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pypdf==5.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-magic==0.4.27
python-magic-bin==0.4.14
python-multipart==0.0.12
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.1.1
tokenizers==0.20.0
torch==2.4.1
tqdm==4.66.5
transformers==4.45.2
